+++ frmVendedor.cs
@@ -20,18 +20,46 @@ namespace pryBaldovinoSP1EPR
 
         private void cmdRegistrar2_Click(object sender, EventArgs e)
         {
-            StreamWriter Vendedor = new StreamWriter("./Vendedor.csv", true);
-            Vendedor.WriteLine(txtIdentificacion2.Text + ";" + txtNombre2.Text);
-            MessageBox.Show("Archivo creado con éxito.");
-            Vendedor.Close();
-            txtIdentificacion2.Text = "";
-            txtNombre2.Text = "";
-            txtIdentificacion2.Focus();
+            bool repetido = false;
+            if (File.Exists("./Vendedor.csv"))
+            {
+                StreamReader ControlarRepVendedor = new StreamReader("./Vendedor.csv");
+                while (!ControlarRepVendedor.EndOfStream)
+                {
+                    string repiteVendedor = ControlarRepVendedor.ReadLine();
+                    string[] Vendedor = repiteVendedor.Split(';');
+                    if (txtIdentificacion2.Text == Vendedor[0])
+                    {
+                        MessageBox.Show("ID repetido, vuelva a intentar");
+                        txtIdentificacion2.Text = "";
+                        txtIdentificacion2.Focus();
+                        repetido = true;
+                    }
+                }
+                ControlarRepVendedor.Close();
+
+                if (repetido == false)
+                {
+                    StreamWriter Vendedor = new StreamWriter("./Vendedor.csv", true);
+                    Vendedor.WriteLine(txtIdentificacion2.Text + ";" + txtNombre2.Text + ";" + lstActivo.SelectedItem + ";" + lstComision.SelectedItem);
+                    MessageBox.Show("Datos cargados con éxito.");
+                    Vendedor.Close();
+                    txtIdentificacion2.Text = "";
+                    txtNombre2.Text = "";
+                    txtIdentificacion2.Focus();
+                }
+            }
         }
 
         private void cmdSalir_Click(object sender, EventArgs e)
         {
             this.Close();
+        }
+
+        private void cmdBorrar_Click(object sender, EventArgs e)
+        {
+            File.Delete("./Vendedor.csv");
+            MessageBox.Show("Usted borró el archivo");
         }
     }
 }
